+++ se24_PT04/src/sg/edu/nus/iss/ssa/bo/TotalReceiptCalculator.java
@@ -14,7 +14,7 @@ public class TotalReceiptCalculator {
 
     Order order = null;
     DiscountOfferCalculator offerCalculator = null;
-    Map<Long, Product> productMap;
+    Map<Integer, Product> productMap;
     List<Transaction> transactionList;
 
     public TotalReceiptCalculator(){
